[#45] refactor: type.ts -> types.ts로 네이밍 변경
